requirements.txt
django==4.2.19
djangorestframework==3.15.2
postmarker==1.0